restructure to mv clirr package into java
