+++ internal/librarian/java/clirr.go
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package clirr handles the generation of Clirr ignore files for Java libraries.
-package clirr
+package java
 
 import (
 	"embed"
@@ -39,7 +38,7 @@ const (
 	templateName = "clirr-ignored-differences.xml.tmpl"
 )
 
-// Generate generates the clirr-ignored-differences.xml file if it doesn't exist.
+// GenerateClirr generates the clirr-ignored-differences.xml file if it doesn't exist.
 //
 // It identifies Java packages containing Protobuf-generated code by searching for
 // files ending in "OrBuilder.java" under "src/main/java". The "OrBuilder" suffix
@@ -49,7 +48,7 @@ const (
 // The generated file contains a set of whitelist rules that tell the Clirr tool
 // to ignore specific changes (like method additions to interfaces) to
 // prevent false-positive binary compatibility failures in the build.
-func Generate(protoModulePath string) error {
+func GenerateClirr(protoModulePath string) error {
 	outputPath := filepath.Join(protoModulePath, clirrIgnoreFile)
 	_, err := os.Stat(outputPath)
 	switch {
