+++ internal/librarian/java/template/clirr-ignored-differences.xml.tmpl
